% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNewItems.R
\name{checkNewItems}
\alias{checkNewItems}
\title{Function to check if there are any problems with new items}
\usage{
checkNewItems(item_ids, creationDate = NULL, domain_id = NULL)
}
\arguments{
\item{item_ids}{ids of new items}

\item{creationDate}{date of item creation}

\item{domain_id}{id of domain}
}
\value{
item summary if everythin is ok with new items
}
\description{
Function to check if there are any problems with new items
}
