% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRating.R
\name{convertRating}
\alias{convertRating}
\title{Rating Transformation Function}
\usage{
convertRating(rating, domainId, type = c("rating", "transformed rating",
  "Q-Score"), database = "oefenweb_nl_app", printResults = TRUE)
}
\arguments{
\item{rating}{Theta, transformed rating or Qscore as input rating}

\item{domainId}{Game domain id}

\item{type}{Type of input rating, it should be "rating" for a normal theta rating,
"transformed rating" if the input rating is a transformed rating and "Q-Score" if the input
rating is a Q-Score}

\item{database}{name of database}

\item{printResults}{should results be printed?}
}
\value{
Converted versions of input rating
}
\description{
Rating Transformation Function
}
