% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createDir}
\alias{createDir}
\title{Function to create directories}
\usage{
createDir(mainDir, subDir = NULL)
}
\arguments{
\item{mainDir}{Name of directory to create (in working directory)}

\item{subDir}{Name of sub directory that should be created}
}
\value{
NULL (creates a directory)
}
\description{
Function to create directories
}
