% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_splom.R
\name{custom_splom}
\alias{custom_splom}
\title{Custom scatterplot matrix (SPLOM)}
\usage{
custom_splom(df, ..., pch = 19, col = grey(0.5), cex = 0.8)
}
\arguments{
\item{df}{(data.frame) data frame whose columns are plotted against each other}

\item{...}{other arguments passed to the function}

\item{pch}{(numeric) the plotting symbol to be used}

\item{col}{(character) the color to be used}

\item{cex}{(numeric) character size of the symbol}
}
\description{
Custom wrapper function around lattice splom with different upper and lower panel
}
