% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ExtractDigits}
\alias{ExtractDigits}
\title{Function to extract all digits}
\usage{
ExtractDigits(x)
}
\arguments{
\item{x}{String}
}
\value{
Numbers in in input string
}
\description{
Function to extract all digits
}
