% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNewItems.R
\name{getItemIds}
\alias{getItemIds}
\title{Function to get item ids based on creation date}
\usage{
getItemIds(creationDate, domain_id)
}
\arguments{
\item{creationDate}{date of item creation}

\item{domain_id}{id of domain}
}
\value{
item_ids of relevant items
}
\description{
Function to get item ids based on creation date
}
