% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTopGO.R
\name{GetTopGO}
\alias{GetTopGO}
\title{Convenience wrapper to TopGO package (Rahnenfueher et al.)}
\usage{
GetTopGO(
  df = NULL,
  GeneID = NULL,
  Gene.ontology.IDs = NULL,
  cluster = NULL,
  selected.cluster,
  topNodes = 50
)
}
\arguments{
\item{df}{an (optional) data.frame with the three columns specified below}

\item{GeneID}{(character) The column containing gene IDs, alternatively a vector}

\item{Gene.ontology.IDs}{(character) The column containing a list of GO terms for each gene, 
alternatively a vector with same order and length as 'GeneID'}

\item{cluster}{(numeric, factor, character) the column containing a grouping variable, 
alternatively a vector with same order and length as 'GeneID'}

\item{selected.cluster}{(character) the name of the group that is to be comapred to the background.
Must be one of 'cluster'}

\item{topNodes}{(numeric) the max number of GO terms (nodes) to be returned by the function.}
}
\value{
a data.frame with TOpGO gene enrichment results
}
\description{
This function carries out a TopGO gene ontology enrichment on a data set
with custom protein/gene IDs and GO terms. The function takes as main 
input a data frame with three specific columns, cluster numbers, Gene IDs, 
and GO terms. Alternatively, these can also be supplied as three individual 
lists.
}
