% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_models.R
\name{gompertzm_fun}
\alias{gompertzm_fun}
\title{Modified Gompertz growth model}
\usage{
gompertzm_fun(LOG10N0, LOG10Nmax, mumax, lag, t)
}
\arguments{
\item{LOG10N0}{(numeric) log 10 initial biomass density}

\item{LOG10Nmax}{(numeric) log 10 maximum biomass density}

\item{mumax}{(numeric) maximum specific growth rate}

\item{lag}{(numeric) lag phase in time units}

\item{t}{(numeric) discrete time points}
}
\value{
(numeric) vector of modeled biomass corresponding to input time points
}
\description{
Selection of growth models
}
