% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lintrProfile}
\alias{lintrProfile}
\title{Lint script with Oefenweb lintr profile}
\usage{
lintrProfile(lintrPath)
}
\arguments{
\item{lintrPath}{Full path of to-be-linted file}
}
\value{
Lintr output
}
\description{
Lint script with Oefenweb lintr profile
}
