% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_models.R
\name{OD_corr}
\alias{OD_corr}
\title{Optical density conversion of plate reader to photometer measurements (OD 600 nm)}
\usage{
OD_corr(OD)
}
\arguments{
\item{OD}{optical density measurements in plate reader}
}
\value{
corrected optical density as obtained in photometer
}
\description{
Selection of growth models
}
