% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_panel_functions.R
\name{panel.barplot}
\alias{panel.barplot}
\title{Draw barplot with error bars in lattice plots}
\usage{
panel.barplot(x, y, ewidth = 0.08, FUN_mean = function(x) mean(x, na.rm
  = TRUE), FUN_errb = function(x) sd(x, na.rm = TRUE), ...)
}
\arguments{
\item{x, y}{(numeric, character) variables to be plotted. The x variable 
is treated as a grouping varibale, i.e. error bars are calculated between groups
of unique x values.}

\item{ewidth}{width of the error bar whiskers}

\item{FUN_mean}{the function used to calculate group (x-variable) means}

\item{FUN_errb}{the function used to calculate group (x-variable) errors}

\item{...}{other arguments passed to the function}
}
\description{
Custom panel functions for lattice plots
}
