% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_panel_functions.R
\name{panel.key}
\alias{panel.key}
\title{Draw custom keys in lattice plots}
\usage{
panel.key(labels, which.panel = 1, pch = 1, cex = 0.8,
  point.cex = NULL, points = TRUE, lines = FALSE,
  rectangles = FALSE,
  col = trellis.par.get()$superpose.polygon$col[1:length(labels)],
  lwd = trellis.par.get()$superpose.line$lwd[1],
  lty = trellis.par.get()$superpose.line$lty[1], corner = c(0.1, 0.9),
  ...)
}
\arguments{
\item{labels}{(character) list of the labels to draw}

\item{which.panel}{(numeric) the panel(s) where key(s) should be drawn}

\item{points}{(logical) if points should be drawn}

\item{lines}{(logical) if lines should be drawn}

\item{rectangles}{(logical) if rectangles should be drawn}

\item{col, lwd, lty, pch, cex, point.cex}{graphical parameters to draw key labels
and symbols}

\item{corner}{(numeric) vector of length 2 indicating the position of the key,
in Normalised Parent Coordinates (0 to 1)}

\item{...}{other arguments passed to the function}
}
\description{
Custom panel functions for lattice plots
}
