% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_panel_functions.R
\name{panel.pvalue}
\alias{panel.pvalue}
\title{Calculate and draw p-values in lattice plots}
\usage{
panel.pvalue(x, y, std = NULL, symbol = TRUE, cex_symbol = 1.5,
  offset = 1, fixed_pos = NULL, verbose = FALSE, pval_digits = 4,
  col = trellis.par.get()$superpose.polygon$col[1], ...)
}
\arguments{
\item{x, y}{(numeric, character) variables to be plotted. The x variable 
is treated as a grouping varibale, i.e. p-values are calculated between groups
of unique x values.}

\item{std}{(character, numeric) the value of x that is used as standard. If NULL, the 
first value is used as standard. If a numeric scalar i, std is determined 
as the i'th element of unique(x). If character, it must be one of x. If 
std = 'all_values', each unique group of x is compared to the total population}

\item{symbol}{(logical) if '*' symbols are to be drawn for significance}

\item{cex_symbol}{(numeric) character size of the symbol}

\item{offset}{(numeric) offset added to the the vertical position of the p-value}

\item{fixed_pos}{(numeric) vertical position of the p-value, 
if NULL determined from the data}

\item{verbose}{(logical) if a summary of the p-value calculation should be 
printed to the terminal}

\item{pval_digits}{(numeric) Integer specifying number of digits for p-value}

\item{col}{(character) the color to be used}

\item{...}{other arguments passed to the function}
}
\description{
Custom panel functions for lattice plots
}
