% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_kegg_brite.R
\name{parse_kegg_brite}
\alias{parse_kegg_brite}
\title{Parse Kegg Brite xml files step-by-step}
\usage{
parse_kegg_brite(data)
}
\arguments{
\item{data}{a data frame with 5 columns as input}
}
\description{
This script is a small utility to parse Kegg Brite XML files and return
a regular data frame instead.
}
\details{
Changes that need to be made to the Kegg XML file before applying function (!):
- replace double spaces \code{'  '} by commas \code{','} or tabs \code{'\t'}
- remove first lines until regular content begins
- possibly add some trailing tabs or commas to end of first line
  so that read.table knows how many columns to expect
}
