% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Progressor}
\alias{Progressor}
\title{Track progress of a for loop}
\usage{
Progressor(current, max)
}
\arguments{
\item{current}{Current iteration of for loop}

\item{max}{Maximum iteration of for loop}
}
\value{
Prints out progress in percentage
}
\description{
Track progress of a for loop
}
