% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRating.R
\name{transRating}
\alias{transRating}
\title{Function to transform rating into transformed rating}
\usage{
transRating(theta, beta)
}
\arguments{
\item{theta}{Normal theta rating}

\item{beta}{All item ratings of the associated items}
}
\value{
Transformed rating
}
\description{
Function to transform rating into transformed rating
}
