% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{UpperCase}
\alias{UpperCase}
\title{Function to make first letter of string uppercase}
\usage{
UpperCase(x)
}
\arguments{
\item{x}{String}
}
\value{
Input string with first letter capitalized
}
\description{
Function to make first letter of string uppercase
}
