% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_aggregation.R
\name{aggregate_pep}
\alias{aggregate_pep}
\title{Aggregate peptide abundances to protein abundances}
\usage{
aggregate_pep(data, sample_cols, protein_col, weight_col,
  weight_threshold = 0.5, method = "sum")
}
\arguments{
\item{data}{the input data frame}

\item{sample_cols}{the columns to be used for peptide aggregation}

\item{protein_col}{the column containing unique protein IDs/names}

\item{weight_col}{the column containing weights or covariance scores}

\item{weight_threshold}{Covariance score (weight) cutoff, Diffacto's default is 0.5}

\item{method}{aggregation method, one of ('sum', 'weightedsum', 'mean', 'weightedmean', 'wgeomean'). 
The default is 'sum'}
}
\value{
a data frame with aggregated protein intensities, one protein at a  row
}
\description{
Similar to the openMS module ProteinQuantifier, this function provides different
methods to aggregate peptide intensities to their parent proteins. It is mainly
intended for the use with (raw) Diffacto results, a table of peptide intensities
and covariation scores (weights) that can be used to filter peptides before aggregating
them up to protein abundances.
}
