% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette_analysis.R
\name{silhouetteAnalysis}
\alias{silhouetteAnalysis}
\title{Wrapper function to perform silhouette analysis on all clusters of a hclust object}
\usage{
silhouetteAnalysis(mat, cluster_object, n_clusters)
}
\arguments{
\item{mat}{(numeric matrix) the original data matrix that clustering was performed on}

\item{cluster_object}{(hclust) the cluster object obtained from running hclust()}

\item{n_clusters}{(numeric) a vector of cluster numbers for which silhouette analysis is
performed}
}
\value{
a list with three objects, silhouette analysis data, and two lattice result plots
}
\description{
Silhouette analysis shows the clusters that have explanatory power.
That includes clusters that are best separated from the neighbours (positive score)
}
\details{
this function applies the iterative silhouette analysis
and stores results in a list

Prerequesite for silhouette analysis is a cluster object that can be obtained
by e.g. running hclust(d = dist(mat), method = "ward.D").
}
